Update alg.cpp тип?

+++ src/alg.cpp
@@ -43,7 +43,7 @@ uint64_t sumPrime(uint64_t hbound) {
   return sumn;
 }
 
-uint64_t twinPrimes(int lbound, int hbound) {
+uint64_t twinPrimes(uint64_t lbound, uint64_t hbound) {
   uint64_t icountTwin = 0;
   for (int i = lbound; i < (hbound - 2); i++) {
     if (checkPrime(i) && checkPrime(i + 2)) icountTwin++;
